+++ delphi/program_analysis/parse_dbn_json.py
@@ -18,9 +18,9 @@ def main():
 
     scope_tree = scp.scope_tree_from_json(data)
 
-    g1 = setup_directed_graph()
-    scope_tree.build_linked_graph(g1)
-    g1.write("linked_graph.dot")
+    # g1 = setup_directed_graph()
+    # scope_tree.build_linked_graph(g1)
+    # g1.write("linked_graph.dot")
 
     g2 = setup_directed_graph()
     scope_tree.build_containment_graph(g2)
